[#3435] - hide empty device message when devices are available
